init_lock

+++ include/stdint.h
@@ -9,8 +9,8 @@ typedef long int64_t;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long uint64_t;
+typedef unsigned long long uint64_t;
 
-typedef unsigned long size_t;
+typedef unsigned long long size_t;
 
 #endif
